Fixed unstable tests, updated version, moved default routers to .api
