+++ pom.xml
@@ -9,7 +9,7 @@
   </parent>
 
   <artifactId>scalecube-services-parent</artifactId>
-  <version>2.12.4-SNAPSHOT</version>
+  <version>2.12.5-SNAPSHOT</version>
   <packaging>pom</packaging>
   <name>ScaleCube-Services</name>
