+++ services-api/src/main/java/io/scalecube/services/routing/StaticAddressRouter.java
