+++ services-gateway/src/main/java/io/scalecube/services/gateway/http/HttpGatewayAcceptor.java
@@ -21,8 +21,8 @@ import io.scalecube.services.api.ServiceMessage;
 import io.scalecube.services.exceptions.ServiceException;
 import io.scalecube.services.exceptions.ServiceProviderErrorMapper;
 import io.scalecube.services.gateway.ReferenceCountUtil;
-import io.scalecube.services.gateway.client.StaticAddressRouter;
 import io.scalecube.services.registry.api.ServiceRegistry;
+import io.scalecube.services.routing.StaticAddressRouter;
 import io.scalecube.services.transport.api.DataCodec;
 import java.util.List;
 import java.util.function.BiFunction;
