+++ services-transport-parent/services-transport-rsocket/pom.xml
@@ -4,7 +4,7 @@
   <parent>
     <groupId>io.scalecube</groupId>
     <artifactId>scalecube-services-transport-parent</artifactId>
-    <version>2.12.4-SNAPSHOT</version>
+    <version>2.12.5-SNAPSHOT</version>
   </parent>
 
   <artifactId>scalecube-services-transport-rsocket</artifactId>
