+++ services/pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <artifactId>scalecube-services-parent</artifactId>
     <groupId>io.scalecube</groupId>
-    <version>2.12.4-SNAPSHOT</version>
+    <version>2.12.5-SNAPSHOT</version>
   </parent>
 
   <artifactId>scalecube-services</artifactId>
